Extract watcherstore (previous storedb) into module
