+++ src/walk_watcher/watcherconfig.py
@@ -1,7 +1,30 @@
 from __future__ import annotations
 
 import logging
+import os
 from configparser import ConfigParser
+
+NEW_CONFIG = """\
+[system]
+database_path = {filename}
+max_is_running_seconds = 300
+oldest_directory_row_days = 30
+oldest_file_row_days = 30
+max_files_per_directory = 1000
+
+[watcher]
+# Metric names cannot contain spaces or commas.
+metric_name = file.watcher
+root_directory = .
+remove_prefix = .
+
+# Exclude directories and files from being watched.
+# The following are regular expressions and are matched against the full path.
+# Multiline values are combined into a single regular expression.
+exclude_directories = ^\\..*$
+exclude_files = ^\\..*$
+
+    """
 
 
 class WatcherConfig:
@@ -72,3 +95,15 @@ class WatcherConfig:
         config_line = self._config.get("watcher", "exclude_files", fallback="")
         lines = [line.strip() for line in config_line.splitlines() if line.strip()]
         return "|".join(lines) or None
+
+
+def write_new_config(filename: str) -> None:
+    """Write a new config file if one does not exist."""
+    if os.path.exists(filename):
+        return
+
+    config_name = filename.replace(".ini", ".db")
+    config = NEW_CONFIG.format(filename=config_name)
+
+    with open(filename, "w") as config_file:
+        config_file.write(config)
