+++ tests/watcherstore_test.py
@@ -6,9 +6,9 @@ from unittest.mock import patch
 
 import pytest
 
-from walk_watcher.walkwatcher import Directory
-from walk_watcher.walkwatcher import File
-from walk_watcher.walkwatcher import StoreDB
+from walk_watcher.watcherstore import Directory
+from walk_watcher.watcherstore import File
+from walk_watcher.watcherstore import WatcherStore
 
 MAX_IS_RUNNING_AGE = 5 * 60  # 5 minutes
 
@@ -34,16 +34,16 @@ FILE_ROWS = [
 
 
 @pytest.fixture
-def store_db() -> StoreDB:
-    return StoreDB(
+def store_db() -> WatcherStore:
+    return WatcherStore(
         ":memory:",
         max_is_running_age=MAX_IS_RUNNING_AGE,
     )
 
 
 @pytest.fixture
-def store_db_full() -> StoreDB:
-    db = StoreDB(
+def store_db_full() -> WatcherStore:
+    db = WatcherStore(
         ":memory:",
         max_is_running_age=MAX_IS_RUNNING_AGE,
     )
@@ -66,7 +66,7 @@ def test_storedb_built_from_config() -> None:
         oldest_directory_row_days=2,
         oldest_file_row_days=3,
     )
-    store_db = StoreDB.from_config(config)
+    store_db = WatcherStore.from_config(config)
 
     assert store_db._max_is_running_age == 1
     assert store_db._oldest_directory_row_age == 2
@@ -74,9 +74,9 @@ def test_storedb_built_from_config() -> None:
 
 
 def test_storedb_works_with_context_manager() -> None:
-    with patch("walk_watcher.walkwatcher.StoreDB.start_run") as start_run_mock:
-        with patch("walk_watcher.walkwatcher.StoreDB.stop_run") as stop_run_mock:
-            with StoreDB(":memory:") as store_db:
+    with patch("walk_watcher.watcherstore.WatcherStore.start_run") as start_run_mock:
+        with patch("walk_watcher.watcherstore.WatcherStore.stop_run") as stop_run_mock:
+            with WatcherStore(":memory:") as store_db:
                 store_db.start_run()
                 store_db.stop_run()
 
@@ -84,7 +84,7 @@ def test_storedb_works_with_context_manager() -> None:
     assert stop_run_mock.call_count == 2
 
 
-def test_create_file_table(store_db: StoreDB) -> None:
+def test_create_file_table(store_db: WatcherStore) -> None:
     cursor = store_db._connection.cursor()
     cursor.execute("SELECT name FROM sqlite_master WHERE type='table'")
     tables = [name[0] for name in cursor.fetchall()]
@@ -92,7 +92,7 @@ def test_create_file_table(store_db: StoreDB) -> None:
     assert "files" in tables
 
 
-def test_create_directory_table(store_db: StoreDB) -> None:
+def test_create_directory_table(store_db: WatcherStore) -> None:
     cursor = store_db._connection.cursor()
     cursor.execute("SELECT name FROM sqlite_master WHERE type='table'")
     tables = [name[0] for name in cursor.fetchall()]
@@ -100,7 +100,7 @@ def test_create_directory_table(store_db: StoreDB) -> None:
     assert "directories" in tables
 
 
-def test_create_system_table(store_db: StoreDB) -> None:
+def test_create_system_table(store_db: WatcherStore) -> None:
     cursor = store_db._connection.cursor()
     cursor.execute("SELECT name FROM sqlite_master WHERE type='table'")
     tables = [name[0] for name in cursor.fetchall()]
@@ -108,7 +108,7 @@ def test_create_system_table(store_db: StoreDB) -> None:
     assert "system" in tables
 
 
-def test_system_table_is_saved_on_init(store_db: StoreDB) -> None:
+def test_system_table_is_saved_on_init(store_db: WatcherStore) -> None:
     cursor = store_db._connection.cursor()
     cursor.execute("SELECT * FROM system")
     rows = cursor.fetchall()
@@ -119,7 +119,7 @@ def test_system_table_is_saved_on_init(store_db: StoreDB) -> None:
     # We don't check the last_seen timestamp for simplicity.
 
 
-def test_start_run_updates_system_table(store_db: StoreDB) -> None:
+def test_start_run_updates_system_table(store_db: WatcherStore) -> None:
     store_db.start_run()
     cursor = store_db._connection.cursor()
     cursor.execute("SELECT is_running FROM system")
@@ -129,7 +129,7 @@ def test_start_run_updates_system_table(store_db: StoreDB) -> None:
 
 
 def test_start_run_with_stale_is_running_flag_updates_system_table(
-    store_db: StoreDB,
+    store_db: WatcherStore,
 ) -> None:
     store_db.start_run()
     cursor = store_db._connection.cursor()
@@ -147,14 +147,14 @@ def test_start_run_with_stale_is_running_flag_updates_system_table(
     assert rows[0] != new_timestamp  # last_seen was updated
 
 
-def test_start_run_while_running_raises(store_db: StoreDB) -> None:
+def test_start_run_while_running_raises(store_db: WatcherStore) -> None:
     store_db.start_run()
 
     with pytest.raises(RuntimeError):
         store_db.start_run()
 
 
-def test_end_run_updates_system_table(store_db: StoreDB) -> None:
+def test_end_run_updates_system_table(store_db: WatcherStore) -> None:
     cursor = store_db._connection.cursor()
     cursor.execute("UPDATE system SET is_running = 1")
 
@@ -165,7 +165,7 @@ def test_end_run_updates_system_table(store_db: StoreDB) -> None:
     assert rows[0] == 0  # is_running is False
 
 
-def test_save_directories(store_db: StoreDB) -> None:
+def test_save_directories(store_db: WatcherStore) -> None:
     directories = [
         Directory("/home/user/magamind", 1618224000, 15),
         Directory("/home/user/minion", 1618224000, 10),
@@ -185,7 +185,7 @@ def test_save_directories(store_db: StoreDB) -> None:
     )
 
 
-def test_save_files_empty_rows(store_db: StoreDB) -> None:
+def test_save_files_empty_rows(store_db: WatcherStore) -> None:
     files = [
         File("/home/user/magamind", "file1", 1618224000),
         File("/home/user/magamind", "file2", 1618224000),
@@ -211,7 +211,7 @@ def test_save_files_empty_rows(store_db: StoreDB) -> None:
     )
 
 
-def test_save_file_existing_row_updated(store_db: StoreDB) -> None:
+def test_save_file_existing_row_updated(store_db: WatcherStore) -> None:
     files = [
         File("/home/user/magamind", "file1", 1618224000),
         File("/home/user/magamind", "file2", 1618224000),
@@ -239,7 +239,7 @@ def test_save_file_existing_row_updated(store_db: StoreDB) -> None:
     assert last_row == (files[-1].last_seen, new_age)
 
 
-def test_save_file_add_new_row_with_existing_rows(store_db: StoreDB) -> None:
+def test_save_file_add_new_row_with_existing_rows(store_db: WatcherStore) -> None:
     files = [
         File("/home/user/magamind", "file1", 1618224000),
         File("/home/user/magamind", "file2", 1618224000),
@@ -267,7 +267,7 @@ def test_save_file_add_new_row_with_existing_rows(store_db: StoreDB) -> None:
     )
 
 
-def test_save_files_marks_all_removed(store_db: StoreDB) -> None:
+def test_save_files_marks_all_removed(store_db: WatcherStore) -> None:
     files = [
         File("/home/user/magamind", "file1", 1618224000),
         File("/home/user/magamind", "file2", 1618224000),
@@ -285,14 +285,14 @@ def test_save_files_marks_all_removed(store_db: StoreDB) -> None:
     assert all(removed)
 
 
-def test_get_directory_rows(store_db_full: StoreDB) -> None:
+def test_get_directory_rows(store_db_full: WatcherStore) -> None:
     rows = store_db_full.get_directory_rows()
 
     # Reference DIRECTORY_ROWS for the expected result
     assert len(rows) == 4
 
 
-def test_get_file_rows(store_db_full: StoreDB) -> None:
+def test_get_file_rows(store_db_full: WatcherStore) -> None:
     directory = Directory(*DIRECTORY_ROWS[0][1:])  # type: ignore
     expected_count = DIRECTORY_ROWS[0][3]
 
@@ -301,7 +301,7 @@ def test_get_file_rows(store_db_full: StoreDB) -> None:
     assert len(rows) == expected_count
 
 
-def test_get_oldest_files(store_db_full: StoreDB) -> None:
+def test_get_oldest_files(store_db_full: WatcherStore) -> None:
     rows = store_db_full.get_oldest_files()
 
     # Reference FILE_ROWS for the expected result
